+++ src/qtmultimediaquicktools/qdeclarativevideooutput_window.cpp
