+++ src/qtmultimediaquicktools/qtmultimediaquicktools.pro
@@ -11,10 +11,24 @@ DEFINES += QT_BUILD_QTMM_QUICK_LIB
 INCLUDEPATH += ../multimedia/qtmultimediaquicktools_headers/
 
 PRIVATE_HEADERS += \
+    ../multimedia/qtmultimediaquicktools_headers/qdeclarativevideooutput_p.h \
+    ../multimedia/qtmultimediaquicktools_headers/qdeclarativevideooutput_backend_p.h \
     ../multimedia/qtmultimediaquicktools_headers/qsgvideonode_p.h \
     ../multimedia/qtmultimediaquicktools_headers/qtmultimediaquickdefs_p.h
 
 SOURCES += \
-    qsgvideonode_p.cpp
+    qsgvideonode_p.cpp \
+    qdeclarativevideooutput.cpp \
+    qdeclarativevideooutput_render.cpp \
+    qdeclarativevideooutput_window.cpp \
+    qsgvideonode_i420.cpp \
+    qsgvideonode_rgb.cpp \
+    qsgvideonode_texture.cpp
 
-HEADERS += $$PRIVATE_HEADERS
+HEADERS += \
+    $$PRIVATE_HEADERS \
+    qdeclarativevideooutput_render_p.h \
+    qdeclarativevideooutput_window_p.h \
+    qsgvideonode_i420.h \
+    qsgvideonode_rgb.h \
+    qsgvideonode_texture.h
