+++ tests/auto/integration/qdeclarativevideooutput_window/qdeclarativevideooutput_window.pro
@@ -3,9 +3,6 @@ TARGET = tst_qdeclarativevideooutput_window
 QT += multimedia-private qml testlib quick
 CONFIG += testcase
 
-OTHER_FILES += \
-        ../../../../src/imports/multimedia/qdeclarativevideooutput_p.h
-
 SOURCES += \
         tst_qdeclarativevideooutput_window.cpp
